+++ plugins/services/src/js/reducers/JSONParser.js
@@ -11,7 +11,7 @@ import {JSONParser as portMappings} from './serviceForm/PortMappings';
 import {JSONParser as residency} from './serviceForm/Residency';
 import {JSONParser as network} from './serviceForm/Network';
 import {JSONParser as multiContainerNetwork} from './serviceForm/MultiContainerNetwork';
-import {JSONParser as volumeMounts} from './serviceForm/VolumeMounts';
+import {JSONParser as volumeMounts} from './serviceForm/MultiContainerVolumes';
 import {simpleParser} from '../../../../../src/js/utils/ParserUtil';
 import {JSONParser as containers} from './serviceForm/Containers';
