+++ plugins/services/src/js/reducers/JSONReducers.js
@@ -7,7 +7,7 @@ import {JSONReducer as healthChecks} from './serviceForm/HealthChecks';
 import {JSONReducer as labels} from './serviceForm/Labels';
 import {JSONReducer as portDefinitions} from './serviceForm/PortDefinitions';
 import {JSONReducer as residency} from './serviceForm/Residency';
-import {JSONReducer as volumes} from './serviceForm/VolumeMounts';
+import {JSONReducer as volumes} from './serviceForm/MultiContainerVolumes';
 import {JSONReducer as network} from './serviceForm/MultiContainerNetwork';
 import {JSONReducer as ipAddress} from './serviceForm/IpAddress';
 import {JSONReducer as requirePorts} from './serviceForm/RequirePorts';
